+++ examples/hosts/app-integration/external-data/src/model/appModel.ts
@@ -6,7 +6,7 @@
 import { TypedEventEmitter } from "@fluidframework/common-utils";
 import { IContainer } from "@fluidframework/container-definitions";
 import { IContainerRuntime } from "@fluidframework/container-runtime-definitions";
-import { IFluidResolvedUrl } from "@fluidframework/driver-definitions";
+import { IFluidResolvedUrl, IResolvedUrl } from "@fluidframework/driver-definitions";
 
 import type { IAppModel, IAppModelEvents, IBaseDocument } from "../model-interface";
 
@@ -36,7 +36,7 @@ export class AppModel extends TypedEventEmitter<IAppModelEvents> implements IApp
 	/**
 	 * {@inheritDoc IAppModel.containerResolvedUrl}
 	 */
-	public readonly containerResolvedUrl = (): IFluidResolvedUrl => {
-		return this.container?.resolvedUrl as IFluidResolvedUrl;
+	public readonly getContainerResolvedUrl = (): IResolvedUrl | undefined=> {
+		return this.container?.resolvedUrl;
 	};
 }
